+++ CHANGES.md
@@ -2,6 +2,8 @@ TBA
 ---
 
 * Feature: Autocreate (auto create options at parse time, making assumptions)
+* Feature: When parsing options as arrays, push multiple arguments into a
+  single array
 
 1.6.1 (2011-06-01)
 ------------------